refactor: implement CLI commands for MVC component generation and restructure command modules
