+++ tests/unit/test_name_utils.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from flask_mvc.commands.exceptions import InvalidControllerNameError
-from flask_mvc.commands.name_utils import NameUtils
+from flask_mvc.command.exceptions import InvalidControllerNameError
+from flask_mvc.command.name_utils import NameUtils
 
 
 class TestNameUtils:
